refactor: 스크립트 정리
